requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-simple-chatbot==0.0.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
MarkupSafe==2.1.1
nltk==3.7
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.5
redis==4.3.4
regex==2022.10.31
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
